Improve documentation and rename project
