+++ README.md
@@ -1,7 +1,20 @@
-# Hooks Collection
+# Event Listeners Collection
 
-A collection of hooks which define different types framework independent
+A collection of event listeners wrapper, allowing to decouple constraints and preconditions from the listener logic.
 
 ## Requirements
 
 PHP >=7.1
+
+## Listeners Included
+
+### Autoremovable Listeners
+
+**NTimesListener**
+
+Execute a listener N times, remove it after.
+
+**SingleListener**
+
+Execute a listener once.
+
